Development/flask_blog_backend/requirements.txt
alembic==1.17.1
bleach==6.3.0
blinker==1.9.0
boto3==1.42.21
botocore==1.42.21
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.4
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
